LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/content-security-policy/script-src/worker-importscripts.sub-expected.txt


FAIL Dedicated worker inherits CSP from main document assert_equals: expected "importScripts allowed" but got "importScripts blocked"
FAIL Shared worker delivers its own CSP Shared workers are not supported
